requirements.txt
colorlog>=6.7.0
PyQt6>=6.4.0
numpy>=1.24.3
pyautogui>=0.9.54
opencv-python>=4.8.0.74
transitions>=0.9.0
pytesseract==0.3.10
pillow==10.0.0
pyscreeze==0.1.29
python-dateutil==2.8.2
ttkbootstrap==1.10.1
python-dotenv==1.0.0
pynput==1.7.6